perf: got localhost working and json to resolve
