+++ package.json
@@ -5,15 +5,15 @@
   "main": "dist/bundle.js",
   "type": "module",
   "files": [
-    "index-bundle.d.ts",
-    "index-bundle.js",
-    "index-bundle.js.map"
+    "index.d.ts",
+    "bundle.js",
+    "bundle.js.map"
   ],
   "scripts": {
-    "start": "tsc && rollup --config --environment NODE_ENV:development --watch",
-    "build": "tsc && rollup -c --bundleConfigAsCjs",
-    "clean": "rm -rf ./src/bundle.js ./src/bundle.js.map",
-    "deploy": "npm run build && npm run clean && gh-pages -d dist",
+    "start": "tsc && rollup -c -w",
+    "build": "tsc && rollup -c",
+    "deploy": "npm run build && gh-pages -d dist",
+    "serve": "serve",
     "test": "jest"
   },
   "repository": {
@@ -46,6 +46,7 @@
     "gh-pages": "^6.1.1",
     "path": "^0.12.7",
     "rollup": "*",
+    "rollup-plugin-serve": "^1.1.1",
     "rollup-plugin-sourcemaps": "^0.6.3",
     "tslib": "^2.6.2"
   }
