+++ src/components/List/List.ts
@@ -9,10 +9,22 @@ export default class ListComponent extends HTMLElement {
 
     connectedCallback() {
         console.log('ListComponent has mounted to page');
-
-    
     }
 };
+
+export const createListComponent = (data: {[key: string]: object | object[] | string}) => {
+    const list = document.createElement('list-component');
+
+    const menuValues = Object.values(data.sections);
+    menuValues.forEach(item => {
+        const eachListItem = document.createElement('list-item-component')
+        eachListItem.innerHTML = `<p>${item}</p>`;
+        list.appendChild(eachListItem)
+    });
+
+    return list;
+
+}
 
 customElements.define('list-component', ListComponent, {extends: 'ul'})
