+++ src/components/index.ts
@@ -1,10 +1,11 @@
 // each component and helpers
-import ListComponent from "./List/List";
+import ListComponent, {createListComponent} from "./List/List";
 import ListItem from './List/ListItem';
 
 export {
     ListComponent,
     ListItem,
+    createListComponent
 };
 
 export const componentsLoaded = () => console.log('components have loaded');
