+++ src/index.html
@@ -18,7 +18,7 @@
     <meta property="fb:app_id" content="1267381723369445" />
     <link rel="shortcut icon" href="https://res.cloudinary.com/ninjaofawesome/image/upload/v1497362710/littlepioneer/favicon.ico" />
     <title>Little Pioneer Cider House | Music Production NYC</title>
-    <script type="text/javascript" src="./bundle.js" ></script>
+    <script type="text/javascript" src="../dist/bundle.js" ></script>
   </head>
   <body>
     <!-- before deploying, change file name to minified version-->
